projects/financial-analysis-agent/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.12
